The Great Vulkan Migration
Ditched OpenGL and embraced the awesomeness of Vulkan, making the graphics engine more scalable for future improvements.
Removed the dependency on the SFML library and now using GLFW for all windowing and input tasks.
Implemented CSM, MSAA, and fog. Adjusted some colors and added gamma correction for improved visuals.
Copied the Etna library into the project, allowing it to evolve alongside the project and be modified as needed for new Vulkan features.
Removed code related to joystick controls due to the removal of the SFML library.

Apologies for the massive commit. Sometimes I just forget thet version control exists. The old OpenGL version of the project is kept untouched in the last-opengl-version branch